requirements.txt
discord.py>=2.0.0
pytz>=2023.3
python-dotenv>=1.0.0
aiohttp>=3.8.1
psycopg2-binary>=2.9.9
